requirements.txt
antlr4-python3-runtime==4.9.3
anyio==4.11.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
argparse-dataclass==2.0.0
arrow==1.4.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
ax-platform==0.3.6
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.3.0
botorch==0.9.5
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cloudpickle==3.1.2
cmake==4.1.2
comm==0.2.3
configargparse==1.7.1
connection-pool==0.0.3
contourpy==1.3.2
cycler==0.12.1
datrie==0.8.3
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
docutils==0.22.3
dpath==2.2.0
exceptiongroup==1.3.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.0
fonttools==4.60.1
fqdn==1.5.1
gitdb==4.0.12
gitpython==3.1.45
gpytorch==1.11
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
hydra-ax-sweeper==1.1.5
hydra-core==1.3.2
idna==3.11
ipykernel==7.1.0
ipython==8.37.0
ipywidgets==8.1.7
isoduration==20.11.0
jaxtyping==0.3.3
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter-server==2.17.0
jupyter-server-terminals==0.5.3
jupyterlab==4.5.0
jupyterlab-pygments==0.3.0
jupyterlab-server==2.28.0
jupyterlab-widgets==3.0.15
kiwisolver==1.4.9
lark==1.3.1
linear-operator==0.5.1
lit==18.1.8
llvmlite==0.45.1
markupsafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mistune==3.1.4
mpmath==1.3.0
multipledispatch==1.0.0
mypy-extensions==1.1.0
narwhals==2.10.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.5.0
notebook-shim==0.2.4
numba==0.62.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.28.7
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
opt-einsum==3.4.0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
plotly==6.3.1
polars==1.34.0
polars-runtime-32==1.34.0
prometheus-client==0.23.1
prompt-toolkit==3.0.52
psutil==7.1.3
ptyprocess==0.7.0
pulp==2.7.0
pure-eval==0.2.3
pycparser==2.23
pygments==2.19.2
pyparsing==3.2.5
pyre-extensions==0.0.32
pyro-api==0.1.2
pyro-ppl==1.9.1
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
reretry==0.11.8
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.29.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
send2trash==1.8.3
setuptools==80.9.0
shap==0.49.1
six==1.17.0
slicer==0.0.8
smart-open==7.5.0
smmap==5.0.2
snakemake==7.32.4
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
stopit==1.1.2
sympy==1.14.0
tabulate==0.9.0
terminado==0.18.1
threadpoolctl==3.6.0
throttler==1.2.2
tinycss2==1.4.0
tomli==2.3.0
toposort==1.10
torch==2.0.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
triton==2.0.0
typeguard==2.13.3
typing-extensions==4.15.0
typing-inspect==0.9.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
wadler-lindig==0.1.7
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
wheel==0.45.1
widgetsnbextension==4.0.14
wrapt==2.0.1
xgboost==3.1.1
yte==1.9.0
